pop_watcher/requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
idna==2.8
ratelimiter==1.2.0.post0
requests==2.21.0
soupsieve==1.9.1
urllib3==1.24.2
yoyo-migrations==6.1.0